Separate into multiple files and add command line argument parsing
