+++ javascript/src/index.js
@@ -8,7 +8,8 @@ var label_msg = require('./../js_proto/labels_pb.js')
 var output_msg = require('./../js_proto/output_pb.js')
 
 // Formatting
-const num_decimal_places = 3;
+var print_utils = require('./print_utils')
+
 
 
 main()
@@ -16,60 +17,20 @@ main()
 
 function main() {
   
-  //TODO take filename as input
-  let filename = "/home/seoulrobotics-thor/dev/sensr_sdk/javascript/sample_output/0000.bin";
-  parseOutputFile(filename);
+  let args = process.argv.slice(2);
+  if (args === undefined ||  args.length == 0) {
+    console.log("No binary file specified.");
+  } else if (args.length > 1) {
+    console.log("Wrong number of command line arguements.");
+  } else {
 
-}
+    const filename = args[0];
+    parseOutputFile(filename);
 
-
-function labelToString(label) {
-  switch(label) {
-    case label_msg.LabelType.CAR:
-      return "Car";
-    case label_msg.LabelType.PEDESTRIAN:
-      return "Ped";
-    case label_msg.LabelType.CYCLIST:
-      return "Cyclist";
-    case label_msg.LabelType.MISC:
-      return "Misc";
-    default:
-      return "Misc";
   }
-}
 
-function Format(float) {
-  return float.toFixed(num_decimal_places);
-}
+  
 
-function PrettyPrintFloat(float, name) {
-  console.log(`${name}:\n\t${Format(float)}`);
-}
-
-function PrettyPrintVec2(vec, name) {
-  const x = Format(vec.getX());
-  const y = Format(vec.getY());
-  console.log(`${name}:\n\t${x}\n\t${y}`);
-}
-
-function PrettyPrintVec3(vec, name) {
-  const x = Format(vec.getX());
-  const y = Format(vec.getY());
-  const z = Format(vec.getZ());
-  console.log(`${name}:\n\t${x}\n\t${y}\n\t${z}`);
-}
-
-function PrintBbox(bbox) {
-  PrettyPrintVec2(bbox.getPosition(), "position");
-  PrettyPrintVec3(bbox.getSize(), "size");
-  PrettyPrintFloat(bbox.getYaw(), "yaw")
-}
-
-function PrintObject(obj) {
-  const id = obj.getId();
-  const label = labelToString(obj.getLabel());
-  console.log(`Object #${id}: ${label}`);
-  PrintBbox(obj.getBbox());
 }
 
 function parseOutputFile(filename) {
@@ -81,7 +42,7 @@ function parseOutputFile(filename) {
   
   objects.forEach(function(obj, key) {
     console.log("------------------------------");
-    PrintObject(obj);
+    print_utils.printObject(obj);
   });
